solution to part4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})  
-    response.json(blogs)
-  })
+  const blogs = await Blog.find({}).populate('user', { id: 1, name: 1, username: 1}) 
+  response.json(blogs)
+})
   
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   let newBlog = request.body
+  
   if(!Object.keys(newBlog).includes('title') || !Object.keys(newBlog).includes('url')) {
     response.status(400).end()
   }
@@ -18,14 +21,29 @@ blogsRouter.post('/', async (request, response) => {
       likes: 0
     }
   }
-
-  const blog = new Blog(newBlog)
+  
+  const user = request.user
+  const blog = new Blog({
+    ...newBlog,
+    user: user._id
+  })
 
   const savedBlog = await blog.save()
+  user.blogs =  user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id); 
+  if(!blog) {
+    return response.status(404).json({ error: 'blog not found with this id' })
+  }
+  const user = request.user
+  if(blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'user is not authorized to delete this blog' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
